Added dependencies for postgress and h2; Added enum AccountStatus; Added necessary field in POJO clases; Added application.yml file for configuring DataSource for main and tests; Added interface AccountRepository; Added implementation of AccauntRepository; Added RowMapper for Account model; Created schema.sql and data.sql for testing; Added integration tests for AccountRepository;

+++ src/main/java/com/shop/online/simple/entity/Product.java
@@ -9,11 +9,11 @@ public class Product {
     private String name;
     private String description;
     private Set<Tag> tags;
-    private long price;
+    private double price;
     private Set<Photo> photos;
     private Set<Feedback> feedbacks;
 
-    public Product(String name, String description, long price) {
+    public Product(String name, String description, double price) {
         this.name = name;
         this.description = description;
         this.tags = new HashSet<>();
@@ -46,11 +46,11 @@ public class Product {
         this.tags = tags;
     }
 
-    public long getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(long price) {
+    public void setPrice(double price) {
         this.price = price;
     }
